Go back to substring filtering instead of fuzzy filtering
This reverts (more or less) what was done in b46623ebef, but makes it so that
you can search for multiple subwords by separating them with spaces. In my
opinion this gives us the best of both worlds: it provides much more relevant
filter results in general, but still allows to search for multiple substrings if
you don't know exactly how they occur in the item you are looking for.

Additionally, you can switch back to fuzzy filtering on a case-by-case basis by
prefixing the filter string with "~" (meant to resemble the "≈" symbol for
"approximately"). I'm not sure we need to keep this much flexibility, but for
now it makes it easy to do A/B testing of the two mechanisms. (Note that it's
not completely equivalent: if you want to search for discontiguous substrings,
you need to separate them with spaces when substring matching, but you must
_not_ use spaces when fuzzy matching.)