[gocui] Re-render text area of editable views when view size changes
Whenever the size of an editable view changes, try to scroll it to (0,0), and
then make the cursor visible again if needed. This is maybe not the ideal
behavior when the rendered content exceeds the size of the view, but it's good
enough, and simple.

This is relevant in two cases: when the terminal size changes while an editable
view is showing, and the first time we resize a view to its actual size in
layout() after showing it for the first time. For the second case we currently
have code that resizes the panel to its real size before showing the view, so
that the text area is scrolled correctly; we can get rid of that code now (which
is the main purpose of this commit).