[gocui] Cleanup: add empty default cases to silence linter warnings
It annoys me that VS Code shows me yellow warnings for these whenever I edit
gocui files as part of the lazygit project (I'm actually not sure why it does
this, I thought it's not supposed to look into the vendor directory). Anyway,
trying to silence these by adding empty default cases. I'm not sure this is the
best or most idiomatic way to silence the warnings.