[gocui] Work around Ghostty quirk with pasted newlines
When pasting, Ghostty seems to convert all \n to \r for some reason. Convert
them back to \n so that we can handle them normally.

I have to admit that I'm confused as to why this is necessary. From what I
understand from Ghostty's source code (e.g.
https://github.com/ghostty-org/ghostty/commit/010338354a0) it does this
conversion only for non-bracketed paste mode, but I'm seeing it in bracketed
paste mode, and it is only there that we convert it back.