Fix crash when clicking in the status view
The status view is not supposed to be focusable right now. (This might change
soon, but for now it isn't.) Pressing '0' on it does nothing.

However, clicking on it would still focus it, because the click handler in
MainViewController assumed that when the clicked view doesn't have the focus,
then its "other" view must have, and we just want to toggle the focus between
the two (like when pressing tab). It didn't take the possibility into account
that the current side panel isn't focusable at all; if it was, then its
SwitchToFocusedMainViewController would have handled the click.

To fix this, check if the "other" view has the focus before handling the click,
and do nothing otherwise.

This also fixes clicking in the main views of the Worktrees or Submodules tabs,
or any other tabs whose main views are not focusable.